pyproject.toml
[build-system]
requires = [ "setuptools >=65.0", "wheel >=0.36.2",]
build-backend = "setuptools.build_meta"

[project]
name = "app"
dynamic = [ "version", "readme", "license",]
description = "Running an application with a frontend and backend on Modal."
requires-python = ">=3.10"
dependencies = [ "fastapi>=0.88.0", "jinja2>=3.1.2", "modal-client>=0.40.373", "python-dotenv>=0.21.0", "structlog>=22.3.0",]
[[project.authors]]
name = "Anthony Corletti"
email = "anthcor@gmail.com"

[project.optional-dependencies]
test = [ "pytest >=6.2.5", "coverage >=6.1.1", "pytest-cov >=3.0.0",]
dev = [ "black >=21.10b0", "isort >=5.9.3", "mypy >=0.910", "pre-commit >=2.17.0", "ruff >=0.0.98", "uvicorn>=0.20.0",]

[project.urls]
Documentation = "https://github.com/anthonycorletti/modal-fullstack-example"

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
ignore = [ "D10", "D415",]
target-version = "py310"
extend-select = [ "D",]
extend-ignore = [ "D203", "D204", "D213", "D215", "D400", "D404", "D406", "D407", "D408", "D409", "D413",]

[tool.setuptools.dynamic.version]
attr = "app.__version__"
